+++ src/main/java/com/socialmedia/controller/FileHandlerController.java
@@ -1,28 +1,33 @@
 package com.socialmedia.controller;
 
-import com.socialmedia.service.AmazonClient;
+import com.socialmedia.service.AmazonService;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.web.bind.annotation.*;
+
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RequestPart;
+import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.bind.annotation.DeleteMapping;
+import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.multipart.MultipartFile;
 
 @RestController
-@RequestMapping("/storage/")
+@RequestMapping("/api/v1/storage")
 public class FileHandlerController {
+  private AmazonService amazonService;
 
-    private AmazonClient amazonClient;
+  @Autowired
+  FileHandlerController(AmazonService amazonService) {
+    this.amazonService = amazonService;
+  }
 
-    @Autowired
-    FileHandlerController(AmazonClient amazonClient) {
-        this.amazonClient = amazonClient;
-    }
+  @PostMapping("/upload")
+  public String uploadFile(@RequestPart(value = "file") MultipartFile file) {
+    return this.amazonService.uploadFile(file);
+  }
 
-    @PostMapping("/uploadFile")
-    public String uploadFile(@RequestPart(value = "file") MultipartFile file) {
-        return this.amazonClient.uploadFile(file);
-    }
-
-    @DeleteMapping("/deleteFile")
-    public String deleteFile(@RequestPart(value = "url") String fileUrl) {
-        return this.amazonClient.deleteFileFromS3Bucket(fileUrl);
-    }
+  @DeleteMapping("/delete/{fileName}")
+  public boolean deleteFile(@PathVariable(value = "fileName") String fileName) {
+    return this.amazonService.deleteFileFromS3Bucket(fileName);
+  }
 }
