+++ src/main/java/com/socialmedia/dto/chat/ChatDtoIn.java
@@ -6,9 +6,9 @@ import com.socialmedia.dto.user.UserLabelDtoIn;
 import java.util.List;
 
 public class ChatDtoIn {
-    private Long id;
-    private String name;
-    private List<UserLabelDtoIn> participants;
-    private List<ChatMessageDtoIn> messages;
+  private Long id;
+  private String name;
+  private List<UserLabelDtoIn> participants;
+  private List<ChatMessageDtoIn> messages;
 
 }
