+++ src/main/java/com/socialmedia/dto/chat/message/ChatMessageDtoIn.java
@@ -3,10 +3,9 @@ package com.socialmedia.dto.chat.message;
 import com.socialmedia.dto.user.UserLabelDtoIn;
 
 public class ChatMessageDtoIn {
-
-    private Long id;
-    private String text;
-    private Long date;
-    private UserLabelDtoIn author;
+  private Long id;
+  private String text;
+  private Long date;
+  private UserLabelDtoIn author;
 
 }
