+++ src/main/java/com/socialmedia/dto/friend/request/FriendRequestDtoIn.java
@@ -3,7 +3,7 @@ package com.socialmedia.dto.friend.request;
 import com.socialmedia.dto.user.UserLabelDtoIn;
 
 public class FriendRequestDtoIn {
-    private Long id;
-    private Long date;
-    private UserLabelDtoIn requester;
+  private Long id;
+  private Long date;
+  private UserLabelDtoIn requester;
 }
