+++ src/main/java/com/socialmedia/dto/post/PostDtoIn.java
@@ -6,12 +6,12 @@ import com.socialmedia.dto.user.UserLabelDtoIn;
 import java.util.List;
 
 public class PostDtoIn {
-    private Long id;
-    private String message;
-    private Long date;
-    private String image;
-    private Boolean showEveryone;
-    private UserLabelDtoIn author;
-    private List<CommentDtoIn> comments;
-    private List<UserLabelDtoIn> likes;
+  private Long id;
+  private String message;
+  private Long date;
+  private String image;
+  private Boolean showEveryone;
+  private UserLabelDtoIn author;
+  private List<CommentDtoIn> comments;
+  private List<UserLabelDtoIn> likes;
 }
