+++ src/main/java/com/socialmedia/dto/user/UserLabelDtoIn.java
@@ -1,8 +1,8 @@
 package com.socialmedia.dto.user;
 
 public class UserLabelDtoIn {
-    private String username;
-    private String firstName;
-    private String lastName;
-    private String avatar;
+  private String username;
+  private String firstName;
+  private String lastName;
+  private String avatar;
 }
